refactor(backend): moved From instance into saved_tree, allowing for getter removal
