+++ frontend/src/main/scala/model/DebugTree.scala
@@ -11,9 +11,18 @@ import model.json.Reader
   * @param isDebuggable Used for if the tree is being actively used for debugging with breakpoints
   * @param sessionId Id of a debugging session, This will be -1 for trees that cannot be debugged, whereas
   *                  it will be some id for trees that are actively being debugged or have been debugged
+  * @param sessionName The name of the session, which should alter the tab title.
   * @param refs A list of pairs: `Address` and `Reference Value` from `Parsley`'s State
   */
-case class DebugTree(input: String, root: DebugNode, parserInfo: Map[String, List[(Int, Int)]], isDebuggable: Boolean, sessionId: Int, refs: Seq[(Int, String)] = Nil) derives Reader.upickle
+case class DebugTree(
+        input: String,
+        root: DebugNode,
+        parserInfo: Map[String, List[(Int, Int)]],
+        isDebuggable: Boolean,
+        sessionId: Int,
+        sessionName: String,
+        refs: Seq[(Int, String)] = Nil
+    ) derives Reader.upickle
 
 
 /**
